[prompt][xs]: separated prompt function into utils
