+++ lib/init.js
@@ -1,32 +1,14 @@
 const fs = require('fs')
 const path = require('path')
-const prompt = require('prompt')
-const { logger } = require('./utils/log-handler.js')
 const Datapackage = require('datapackage').Datapackage
 const parse = require('csv-parse/lib/sync')
 const infer = require('tableschema').infer
 const urljoin = require('url-join')
-const checkDpIsThere = require('./utils/common.js').checkDpIsThere
 
+const { logger } = require('./utils/log-handler')
+const checkDpIsThere = require('./utils/common').checkDpIsThere
+const prompt = require('./utils/prompt')
 
-/*
- * function to prompt to user
- * @param {Schema} as schema per prompt lib documentation
- * @return {Object} result of user prompt(e.g: name can be accessed by result.name)
- */
-const promptFunction = (schema) => {
-  prompt.message = ''
-  prompt.start()
-  return new Promise((resolve, reject) => {
-    prompt.get(schema, (err, result) => {
-      if (err) {
-        reject(err)
-        logger(err.message+'\n', 'abort', true)
-      }
-      resolve(result)
-    })
-  })
-}
 
 /*
  * function to scan directory
@@ -126,7 +108,7 @@ const shouldAddFiles = async (files, dpObj, currentPath) => {
          }
        }
       }
-      let result = await promptFunction(schemaForFiles)
+      let result = await prompt.promptFunction(schemaForFiles)
       if(result.answer === 'y'){
         let pathForResource = path.join(currentPath, files[i])
         await addResource(pathForResource, dpObj)
@@ -158,7 +140,7 @@ const shouldScanDir = async (dirs, dpObj, currentPath) => {
        }
      }
     }
-    let result = await promptFunction(schemaForDirs)
+    let result = await prompt.promptFunction(schemaForDirs)
 
     if(result.answer === 'y') {
       let nextPath = path.join(currentPath, dirs[j])
@@ -174,17 +156,43 @@ const shouldScanDir = async (dirs, dpObj, currentPath) => {
 }
 
 /*
+* function to ask if it should write datapackage.json
+* @param {descriptor} descriptor of the datapackage
+* does not return anything but stops the process depending on user's input
+*/
+const shouldWrite = async (descriptor) => {
+  const cwd = path.join(process.cwd(), 'datapackage.json')
+  let schemaForShouldWrite = {
+    properties: {
+      answer: {
+        description: `Going to write to ${cwd}:\n\n${JSON.stringify(descriptor, null, 2)} \n\n\nIs that OK - y/n?`,
+        pattern: /^[y,n]+$/,
+        message: `Please, provide with following responses 'y' for yes or 'n' for no`,
+        required: true
+      }
+    }
+  }
+
+  let result = await prompt.promptFunction(schemaForShouldWrite)
+
+  if(result.answer === 'n') {
+    logger(`Process canceled\n`, 'abort', true)
+  }
+}
+
+/*
 * function to write/extend dpObj into datapackage.json
 * @param {dpObj} instance of the datapackage
-* it extends datapackage.json
+* it wrties datapackage.json file to the disk
 */
 const writeDp = async (dpObj) => {
+  const cwd = path.join(process.cwd(), 'datapackage.json')
   const content = JSON.stringify(dpObj._descriptor, null, 2);
   fs.writeFile("./datapackage.json", content, 'utf8', function (err) {
   if (err) {
       return logger(err);
   }
-  logger("datapackage.json file is saved!");
+  logger(`datapackage.json file is saved in ${cwd}`);
   })
 }
 
@@ -201,7 +209,7 @@ const updateDp = async () => {
   let schemaForConfirmationDp = {
    properties: {
      answer: {
-       description: 'There is datapackage.json already. Do you want to update it ' + ' y/n?',
+       description: 'There is datapackage.json already. Do you want to update it - y/n?',
        pattern: /^[y,n]+$/,
        message: `Please, provide with following responses 'y' for yes or 'n' for no`,
        required: true
@@ -209,17 +217,18 @@ const updateDp = async () => {
    }
   }
 
-  let result = await promptFunction(schemaForConfirmationDp)
+  let result = await prompt.promptFunction(schemaForConfirmationDp)
   if(result.answer === 'y'){
     const dpObj = await new Datapackage('datapackage.json')
     let path_ = ''
     let filesAndDirs = await scanDir()
     await shouldAddFiles(filesAndDirs.files, dpObj, path_)
     await shouldScanDir(filesAndDirs.dirs, dpObj, path_)
+    await shouldWrite(dpObj._descriptor)
     await writeDp(dpObj)
   }
   else {
-    logger(`Process cancelled\n`, 'abort', true)
+    logger(`Process canceled\n`, 'abort', true)
   }
 }
 
@@ -241,7 +250,8 @@ const createDp = async () => {
        description: 'Enter Data Package name',
        pattern: /^[a-z\.\-\_]+$/,
        message: `Must consist only of lowercase alphanumeric characters plus ".", "-" and "_"`,
-       required: true
+       default: 'scratchpad',
+       required: false
      },
      title: {
        description: 'Enter Data Package title',
@@ -249,7 +259,7 @@ const createDp = async () => {
      }
    }
   }
-  let result = await promptFunction(schemaForNameAndTitle)
+  let result = await prompt.promptFunction(schemaForNameAndTitle)
 
   let descriptor = {
       name: result.name,
@@ -261,6 +271,7 @@ const createDp = async () => {
   let filesAndDirs = await scanDir()
   await shouldAddFiles(filesAndDirs.files, dpObj, path_)
   await shouldScanDir(filesAndDirs.dirs, dpObj, path_)
+  await shouldWrite(dpObj._descriptor)
   await writeDp(dpObj)
 }
 
@@ -278,11 +289,11 @@ const init = async ()=> {
 }
 
 module.exports.init = init
-module.exports.promptFunction = promptFunction
 module.exports.scanDir = scanDir
 module.exports.addResource = addResource
 module.exports.buildSchema = buildSchema
 module.exports.shouldAddFiles = shouldAddFiles
 module.exports.shouldScanDir = shouldScanDir
+module.exports.shouldWrite = shouldWrite
 module.exports.writeDp = writeDp
 module.exports.createDp = createDp
